third_party/blink/web_tests/fast/forms/number/number-lossless-localization-expected.txt

PASS numberInput.value is "0000123456789012345678901234567890123456789"
PASS successfullyParsed is true

TEST COMPLETE
